refactor: add Provider class in service package
