+++ packages/core/src/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './helpers';
-export * from './providers';
