+++ packages/services/src/index.ts
@@ -1,6 +1,6 @@
-export * from './types';
-export * from './epcisEvents';
-export * from './utils/helpers';
-export * from './utils/httpService';
-export * from './providers';
-export * from './createLinkResolver.service';
+export * from './types/types.js';
+export * from './epcisEvents/index.js';
+export * from './utils/helpers.js';
+export * from './utils/httpService.js';
+export * from './providers/index.js';
+export * from './createLinkResolver.service.js';
