+++ packages/services/src/providers/index.ts
@@ -1,2 +1,3 @@
-export * from './gs1';
-export * from './ProviderStrategy';
+export * from './gs1Provider.js';
+export * from './ProviderStrategy.js';
+export * from './provider.js';
