+++ packages/services/tsconfig.json
@@ -17,7 +17,6 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true,
     "outDir": "build",
     "declaration": true,
     "sourceMap": true,
